+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Update `compute_best_second_guesses` to have more consistent printout
+
 ## [1.4.0] - 2026-03-10
 
 ### Added